fix(studio): address PR review — Fast Refresh + trace bar a11y
- Extract metricHeadline/formatMetricValue from instruments-table into a pure
  instrument-format module, so the component file exports only its component
  (React Doctor only-export-components) — mirrors traces-panel/trace-geometry.
- Give the trace waterfall bar gridcell an aria-label with the span's offset and
  duration; it carried no text, so a screen reader announced an empty cell
  (React Doctor control-has-associated-label).
